fixed issues with training

+++ Transport.cpp
@@ -143,7 +143,7 @@ void Transport::changeDirection() {
 
     int index = this->getProgress() + 1;
 
-    if (index < this->getRoute().size()) {
+    if (index < (int)this->getRoute().size()) {
         this->setDirection(this->getRoute()[index]);
         ENSURE(this->getDirection() == this->getRoute()[index], "chanceDirection did not change direction correctly.");
     } else {
